servo_drive_demo: add linker_mem_map.cmd for AM64x
Common file for AM64x defining the memory in the device. Memory sections
created for R5Fs, M4F, on-chip SRAM, and DDR.

This is a starting point for the linker command file as there currently isn't
any way to test without an AM64x test case to run on a simulator. This file
will evolve as AM64x support is added.

Signed-off-by: Jason Reeder <jreeder@ti.com>