servo_drive_demo: add C2000 slave pre-built binary
Adds the pre-built binary needed to run on the C2000 slave device in the
multi-axis demo.

Due to limitations in the AM65x PRU FSI emulation, the default C2000 slave code
needs to be modified. The accompanying README file explains the process
needed to rebuild an equivalent binary from source given C2000's existing
documentation. C2000 source files with modifications are also included in the
directory to rebuild the slave binary from source.

Signed-off-by: Jason Reeder <jreeder@ti.com>