servo_drive_demo: upgrade to BIOS 6_82 to prevent improper preemption of R5F interrupts
CSS load and run crashes with this version as it expect vectors to be
placed in ATCM (0x0 address). This is an ARM requirement when TCM is
enabled. Added a PATCH (that copy Vectors into ATCM), aligned the same
with PSW.  With this patch sitara apps works along with bios_6_82_00_16
& xdctools_3_61_00_16_core
Minor update of servo drive demo memory map to move code/data to
internal (MSMC) memory for both R5F0 and R5F1

JIRA: SYSBIOS-1218