ipc: implement IPC RPMsg_char on R5
IPC RPMsg_char driver provides a way for the Linux user space program
to communicate with the application program run on R5 cores. It
provides RPMsg channels between the cores (R5 <-> R5, A53 <-> R5). On the
Linux side, the IPC RPMsg_char has been integrated into the Linux kernel.
For R5 side, we need to implement the IPC RPMsg_char protocol as library.
This library code is for that purpose.

JIRA: SITSW-29

Signed-off-by: Ming Wei <mwei@ti.com>