benchmark_demo: add JSON file parsing code
JSON file is widely used for describing the system resources, structured
information etc. We need to pass structured information between the webserver
and the benchmark data providers (R5 cores), naturally the JSON file format
was picked. To make the communication scheme work, we will need to be able to
parse/update a JSON file, so that we can read/write certain fields in
JSON file in real time. JSMN is a world fastest and minimalistic JSON
parser/tokenizer and it is freely available on github.com/zserge/jsmn.

JIRA: SITSW-37

Signed-off-by: Ming Wei <mwei@ti.com>