benchmark_demo: add the IPC RPMsg_char and JSON file updating code on A53
This is a Linux user space program. It uses IPC RPMsg_char API to communicate with
the R5 cores. It sends control commands (what benchmark to run, how often the
benchmark will run) to the R5 and gets the benchmark data
from R5 cores. It will then use the obtained benchmark data to update the JSON file,
so that the webserver can display the updated information in the GUI
interface in real time.

JIRA: SITSW-29, SITSW-37

Signed-off-by: Ming Wei <mwei@ti.com>