sitara-apps: move device specific files into device specific folders
As more devices are added to the SDK it has become clear that a better method
of supporting multiple devices is needed. #ifdef DEVICE_NAME strategy is
actively being avoided in order to avoid bloating files with unnecessary code.
Previously, the solution was device names being post-fixed onto file names which
solved part of the problem but still left an issue when a source file needed to
include a device specific header file.

Current solution being committed creates device specific folders at the src and
include levels of each library and executable to house soc specific files
with no need for post-fixed file names. The $(SITARA_DEMO_SOC) variable in the
concerto.mak file is used to choose the correct include folder and source
folder.

Signed-off-by: Jason Reeder <jreeder@ti.com>