safety_app: allow M4F build to use common linker_mem_map.cmd file
Attempting to use the common/config/am64x/mem_map/linker_mem_map.cmd file
created linking issues due to overlap in the MEMORY sections of the R5F and M4F
local memories. Solution being committed relies on an update to the concerto
repository that defines the core being built for and passes that information to
the linker. With this information the overlapping local memories of the R5F and
M4F can be filtered out based on the core definition. This allows a single
linker_mem_map.cmd file to contain all MEMORY section definitions for the
entire device.

Signed-off-by: Jason Reeder <jreeder@ti.com>