benchmark_demo: R5F Core Loading and Reloading from GUI
When receiving application/option input from the user on the GUI Composer app,
the JSON file will be updated and these updates will need to be applied to the
R5F cores. In the JSON file, there is an input section for each core that
corresponds to "application", "option", and "changed". Each will have a
numerical value which corresponds to the following:
   "application" 0: None, 1: CFFT, 2: FIR, 3: FOC, 4: Speed/Pos loop, 5: ADC/PWM
   "option" (CFFT) 0: None, 1: 128pt, 2: 256pt, 3: 512pt, 4: 1024pt
   "option" (FIR) 0: None, 1: 1Khz, 2: 2Khz, 3: 4Khz, 4: 8Khz
   "option" (FOC, Speed/Pos, ADC/PWM) 0: None, 1: 8Khz, 2: 16Khz, 3: 32Khz, 4: 50Khz
   "changed" 0: no change, 1: changed
When the "changed" value is 1, the application and option inputs should be loaded/reloaded into the corresponding R5F core

JIRA: SITSW-169

Signed-off-by: Ming Wei <mwei@ti.com>