servo_drive_demo: pslctrl: fixup the linking for SBL boot
Update the entry point flag to be the actual reset vectors instead of _c_int00.
Reset vectors will jump directly to _c_int00 if no exception is detected so
functionality remains the same.

Increases the alignment for the .vecs section from 8 to 256 becaues the SYSFW
checks that the value passed as the entry point to be placed into the BOOT_VECT
registers has 8 trailing zeros. Without this change the SBL will fail when
requesting to set the entry point from the SYSFW.

Move the R5F global address linking to use the newly created R5F local address
from the linker MEMORY sections.

Update the R5F MPU settings to add the local BTCM address and open the full
OCSRAM as executable/cacheable memory.

SITSW-180

Signed-off-by: Jason Reeder <jreeder@ti.com>