examples: timesync: unit_test: use sciclient/SYSFW to configure the Compare Event Router
Remove the CSL functions calls that were directly configuring the Compare Event
Router registers and replace them with the appropriate sciclient function calls
to allow SYSFW to configure the router.

Due to firewall additions the cores can no longer write directly to the Compare
Event Router registers. The correct way to configure the router is through the
SYSFW in the DMSC core using sciclient API calls. This commit makes use of the
Sciclient_rmIrqSetRaw function call to directly configure the inputs and the
outputs of the router similar to the prior CSL calls.

SITSW-235

Signed-off-by: Jason Reeder <jreeder@ti.com>