servo_drive_demo: pslctrl: remove UART config from Board_init call
Adds the BOARD_ENABLE definition back to the pslctrl application and removes
the UART configuration request from the call. In the pre-silicon environment
there is a resource conflict between the R5F and A53/Linux when trying to
use UART0. Removing the UART config from the R5F Board_init call fixes the
conflict and allows Linux to continue using UART0 for it's prompt.

SITSW-346

Signed-off-by: Jason Reeder <jreeder@ti.com>