servo_drive_demo: fix the tisci_host_id security context
Fix the tisci_host_id to match the non-secure context in the resource
management configurations. Originally both the application and the
configurations were using the secure host IDs but now the default RM config
is using the correct non-secure ID. This commit brings the application up to
speed with the non-secure IDs.

Signed-off-by: Jason Reeder <jreeder@ti.com>