servo_drive_demo: configure the FSI pins to be brought to the FSI header on the AM64x EVM
The FSI TX0 and RX0 pins are steered both to the FSI header as well as the
high-density application connector on the AM64x EVM. There is a FET that acts
as a 2:1 mux on the board to steer the pins. The select signal for this FET is
the output of an I2C I/O expander. This commit adds the I2C commands to set the
necessary I/O expander port/pin as an output with value logic HIGH in order to
steer the FSI signals to the FSI header.

Signed-off-by: Jason Reeder <jreeder@ti.com>